# Much more efficient parsing of really big `BigInt` values